Fix type mismatch.

+++ src/SubsetSum.CommandLine/SubsetSumOptions.cs
@@ -8,7 +8,7 @@ namespace SubsetSum.CommandLine
         [Option('v', "verbose", Default = false, Required = false, HelpText = "Prints all messages to standard output.")]
         public bool IsVerbose { get; set; }
 
-        [Option('c', "max-concurrency", Default = 1, Required = false, HelpText = "Defines maximum number of concurrent tasks running.")]
+        [Option('c', "max-concurrency", Default = (uint)1, Required = false, HelpText = "Defines maximum number of concurrent tasks running.")]
         public uint MaxConcurrency { get; set; }
 
         [Value(0, Required = true, HelpText = "Sum that is required to achieve.", MetaName = "Sum")]
